feat:ReturnAccessRights - Client Lib Method(Am-290)
