+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/AccessManagementRepository.java
@@ -7,6 +7,7 @@ import org.jdbi.v3.sqlobject.customizer.BindList;
 import org.jdbi.v3.sqlobject.statement.SqlQuery;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import uk.gov.hmcts.reform.amlib.enums.AccessType;
+import uk.gov.hmcts.reform.amlib.enums.AccessorType;
 import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
 import uk.gov.hmcts.reform.amlib.internal.models.ExplicitAccessRecord;
 import uk.gov.hmcts.reform.amlib.internal.models.Role;
@@ -65,4 +66,8 @@ public interface AccessManagementRepository {
         + " where default_perms.role_name in (<userRoles>) and default_perms.permissions & 1 = 1 and default_perms.attribute = '' and cast(resource.default_security_classification as text) in (<securityClassifications>)")
     @RegisterConstructorMapper(ResourceDefinition.class)
     Set<ResourceDefinition> getResourceDefinitionsWithRootCreatePermission(@BindList Set<String> userRoles, @BindList Set<SecurityClassification> securityClassifications);
+
+    @SqlQuery("select * from access_management where resource_id=? and resource_name=? and resource_type =? and cast(accessor_type as text)  = ? and attribute ='' ")
+    @RegisterConstructorMapper(ExplicitAccessRecord.class)
+    List<ExplicitAccessRecord> getExplicitAccessForResource(String resourceId, String resourceName, String resourceType, AccessorType accessorType);
 }
