+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/models/AccessorListByResourceEnvelope.java
@@ -1,0 +1,17 @@
+package uk.gov.hmcts.reform.amlib.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Builder
+public class AccessorListByResourceEnvelope {
+
+    @JsonProperty("explicitAccess")
+    List<AccessResourceEnvelope> explicitAccessResourceEnvelopesList;
+
+    String resourceId;
+}
