+++ am-lib/src/test/java/uk/gov/hmcts/reform/amlib/AccessManagementServiceValidationTest.java
@@ -1,16 +1,12 @@
 package uk.gov.hmcts.reform.amlib;
 
-import com.fasterxml.jackson.core.JsonPointer;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
-import uk.gov.hmcts.reform.amlib.enums.SecurityClassification;
 import uk.gov.hmcts.reform.amlib.helpers.InvalidArgumentsProvider;
 import uk.gov.hmcts.reform.amlib.models.ExplicitAccessGrant;
 import uk.gov.hmcts.reform.amlib.models.ExplicitAccessMetadata;
-import uk.gov.hmcts.reform.amlib.models.Resource;
 import uk.gov.hmcts.reform.amlib.models.ResourceDefinition;
 
-import java.util.Map;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
@@ -53,29 +49,6 @@ class AccessManagementServiceValidationTest {
                 "accessMetadata.resourceDefinition.resourceType - must not be blank",
                 "accessMetadata.resourceDefinition.resourceName - must not be blank",
                 "accessMetadata.attribute - must not be null"
-            ));
-    }
-
-    @ParameterizedTest
-    @ArgumentsSource(InvalidArgumentsProvider.class)
-    void filterResourceMethodShouldRejectInvalidArguments(
-        String userId, Set<String> userRoles, Resource resource,
-        Map<JsonPointer, SecurityClassification> attributeSecurityClassifications) {
-        assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> service.filterResource(userId, userRoles, resource, attributeSecurityClassifications))
-            .withMessageMatching(expectedValidationMessagesRegex(
-                "userId - must not be blank",
-                "userRoles - must not be empty",
-                "userRoles\\[\\].<iterable element> - must not be blank",
-                "resource - must not be null",
-                "resource.id - must not be blank",
-                "resource.definition - must not be null",
-                "resource.definition.serviceName - must not be blank",
-                "resource.definition.resourceType - must not be blank",
-                "resource.definition.resourceName - must not be blank",
-                "resource.data - must not be null",
-                "attributeSecurityClassifications - must not be empty",
-                "attributeSecurityClassifications - must contain root attribute"
             ));
     }
